refactor good, pending update tests
