--- internal/core/models/associated_user_device_template.go
@@ -1,9 +1,0 @@
-package models
-
-type AssociatedUserDeviceTemplate struct {
-	IsTemplateUpdated bool
-	PidURL            string
-	DeviceSettingURL  string
-	DbcURL            string
-	Version           string
-}
